Further refactoring on runtime Lib ITEMs.

+++ LcsBaseStation/main.cpp
@@ -161,7 +161,7 @@ uint8_t bsRefreshActiveSessionCallback( void *uData ) {
 //
 //----------------------------------------------------------------------------------------
 uint8_t lcsReqCallback( uint16_t npId, 
-                        uint8_t item, 
+                        uint16_t item, 
                         uint16_t *arg1, 
                         uint16_t *arg2, 
                         void *uData ) {
@@ -183,7 +183,7 @@ uint8_t lcsReqCallback( uint16_t npId,
 //
 //----------------------------------------------------------------------------------------
 uint8_t lcsRepCallback( uint16_t npId, 
-                        uint8_t item,
+                        uint16_t item,
                         uint16_t arg1, 
                         uint16_t arg2, 
                         uint8_t ret,
