+++ LcsLibraries/LcsDrvOccDetectLib/LcsDrvOccDetectLib.h
@@ -49,7 +49,7 @@ enum LcsDrvOccDetectItems : uint8_t {
 //
 ///---------------------------------------------------------------------------------------
 uint8_t lcsDrvOccDetect( uint16_t boardId, 
-                         uint8_t  item, 
+                         uint16_t  item, 
                          uint16_t *arg1, 
                          uint16_t *arg2,
                          void     *uData );
